ci: add `textlint` configuration and script for markdown linting (#99)
This pull request includes changes to add `textlint` as a new linter for
markdown files. The most important changes include updating the
`package.json` file to include `textlint` and its configuration, as well
as modifying the `sync-client.sh` script to copy the new `textlintrc.js`
configuration file.

### Addition of `textlint`:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R11):
Added a new script `lint:textlint` to run `textlint` on markdown files
and included `textlint` and `textlint-rule-allowed-uris` as
dependencies. The `lint-staged` configuration was updated to run both
`markdownlint` and `textlint` on markdown files.
[[1]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R11)
[[2]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L29-R43)

### Update to configuration scripts:

*
[`sync-client.sh`](diffhunk://#diff-9f218442f37688e98d9dcaf1702d0954aaaabd45dbc04456c4a900bf0f9a24a2R22):
Added a command to copy the `configs/.textlintrc.js` file to the client
directory.