chore: detach `lint-staged` related configs from `package.json` (#108)
This pull request includes changes to the `package.json` and
`sync-client.sh` files. The most important changes involve the removal
of the `lint-staged` configuration from `package.json` and the addition
of a new configuration file in the synchronization script.

Configuration updates:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L33-L43):
Removed the `lint-staged` configuration section which included commands
for `prettier`, `editorconfig-checker`, `eslint`, `markdownlint`, and
`textlint`.

Synchronization script updates:

*
[`sync-client.sh`](diffhunk://#diff-9f218442f37688e98d9dcaf1702d0954aaaabd45dbc04456c4a900bf0f9a24a2R23):
Added a command to copy `configs/lint-staged.config.js` to the current
directory.