chore(*): delete `configs/.eslintrc.js` and update `depedabot.yml` (#165)
This pull request includes changes to the Dependabot configuration and
the ESLint configuration. The most important changes are the addition of
pull request limits in the Dependabot configuration and the removal of
the ESLint configuration file.

Dependabot configuration changes:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R30):
Added `open-pull-requests-limit: 3` to the updates section to limit the
number of open pull requests Dependabot can create.
[[1]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R30)
[[2]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R50)

ESLint configuration changes:

*
[`configs/.eslintrc.js`](diffhunk://#diff-7a77a8dbe1174fa105949878190f97a004d35eeaadfef50ffdcba66c8a78e7b2L1-L20):
Removed the entire ESLint configuration file, which included settings
for root, extends, environment, and parser options.