chore(*): update lint configs and bump package versions (#243)
This pull request includes changes to update the configuration files and
dependencies for the project. The most important changes involve
updating the `prettier` and `eslint` configurations and modifying the
`sync-client.sh` script.

Updates to configuration files and dependencies:

*
[`configs/.prettierrc.cjs`](diffhunk://#diff-2fe8186baf428cc56a09efc43a22edcd9f5252464f546301c0194a8d42eb5bf9L1):
Removed the `prettier-config-bananass` module export.
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L28-R34):
Updated the `eslint-config-bananass` and `prettier-config-bananass`
dependencies to their latest canary versions.

Modifications to `sync-client.sh` script:

*
[`sync-client.sh`](diffhunk://#diff-9f218442f37688e98d9dcaf1702d0954aaaabd45dbc04456c4a900bf0f9a24a2L19-R23):
Removed the copying of `configs/.prettierrc.cjs` and added the copying
of `configs/prettier.config.mjs`.